feat: replace custom popup div with Leaflet native tooltip
- Delete src/ui/mapPopup.js (custom div-based popup with manual coord math)
- Rewrite pins.js _upsertMarker to use marker.bindTooltip / setTooltipContent
- Add _buildTooltipHtml, _capacityNum, _amenityIcons, _escapeHtml helpers
- Rewrite mapPopup.css: reset Leaflet wrapper chrome, keep inner card styles
- Remove MAP_PIN_HOVERED / MAP_PIN_UNHOVERED from events.js (no longer emitted)
- Remove initMapPopup import and call from main.js
