+++ src/main.js
@@ -13,23 +13,22 @@
  *  6. loadGoogleMaps()  → initMap()  — load Maps SDK then mount the map
  *  7. initPins()        — register map pin listeners (needs map ready)
  *  8. initGroupPinLayer() — register group pin overlay listeners
- *  9. initMapPopup()    — register floating map popup listeners
- * 10. addMapControls()  — inject zoom + locate-me buttons into the map
- * 11. initRealtime()    — open Supabase Realtime channel
- * 12. fetchSpots()      — load spots + confidence → dispatch SPOTS_LOADED
- * 13. fetchActiveClaims() — load current claims → dispatch CLAIMS_LOADED
- * 14. Restore selected spot from URL (if ?spot= present)
- * 15. initSmartSuggestions() — wire F1 filter-submission listener
- * 16. initClaim()            — wire F2 claim flow listener
- * 17. initReportFull()       — wire F3 report-full flow listener
- * 18. initGroups()           — wire F4 group create/join listeners
- * 19. initGroupPins()        — wire F5 group pin lifecycle listeners
- * 20. initFilterPanel()      — render + wire filter UI
- * 21. initSidebar() / initBottomSheet() — wire panel controller for viewport
- * 22. Wire URL sync on store events
- * 23. Wire geolocation (request permission, update store on change)
- * 24. Wire MAP_PIN_CLICKED → SELECT_SPOT dispatch
- * 25. Handle ?join= URL param → pre-fill inline join form
+ *  9. addMapControls()  — inject zoom + locate-me buttons into the map
+ * 10. initRealtime()    — open Supabase Realtime channel
+ * 11. fetchSpots()      — load spots + confidence → dispatch SPOTS_LOADED
+ * 12. fetchActiveClaims() — load current claims → dispatch CLAIMS_LOADED
+ * 13. Restore selected spot from URL (if ?spot= present)
+ * 14. initSmartSuggestions() — wire F1 filter-submission listener
+ * 15. initClaim()            — wire F2 claim flow listener
+ * 16. initReportFull()       — wire F3 report-full flow listener
+ * 17. initGroups()           — wire F4 group create/join listeners
+ * 18. initGroupPins()        — wire F5 group pin lifecycle listeners
+ * 19. initFilterPanel()      — render + wire filter UI
+ * 20. initSidebar() / initBottomSheet() — wire panel controller for viewport
+ * 21. Wire URL sync on store events
+ * 22. Wire geolocation (request permission, update store on change)
+ * 23. Wire MAP_PIN_CLICKED → SELECT_SPOT dispatch
+ * 24. Handle ?join= URL param → pre-fill inline join form
  */
 
 // ─── CSS side-effects (Vite bundles these) ────────────────────────────────────
@@ -79,7 +78,6 @@ import { initFilterPanel } from './ui/filterPanel.js';
 import { initSidebar }     from './ui/sidebar.js';
 import { initBottomSheet } from './ui/bottomSheet.js';
 import { showToast }       from './ui/toast.js';
-import { initMapPopup }    from './ui/mapPopup.js';
 
 // ─── Bootstrap ───────────────────────────────────────────────────────────────
 
@@ -112,7 +110,6 @@ async function boot() {
   initPins();
   initGroupPinLayer();
   initMapControls();
-  initMapPopup();
 
   // ── 9. Realtime ───────────────────────────────────────────────────────────────
   subscribeToRealtime();
