+++ src/styles/map.css
@@ -10,7 +10,6 @@
   position: relative;
   width: 100%;
   height: 100%;
-  padding: var(--space-4);
   background-color: var(--color-surface);
   overflow: hidden;
 }
